+++ tests/testthat/test_sim_n_datasets.r
@@ -68,4 +68,27 @@ test_that("with td, parallel", {
                                max_t=100))
 })
 
+test_that("custom function for prob_fun, parallel", {
+
+  fun2 <- function() {
+    return(1)
+  }
+
+  prob_fun <- function(data) {
+    test <- fun2()
+    return(0.1)
+  }
+
+  assign("prob_fun", value=prob_fun, envir=.GlobalEnv)
+  assign("fun2", value=fun2, envir=.GlobalEnv)
+
+  dag <- empty_dag() +
+    node("A", "rnorm", mean=10, sd=120) +
+    node("B", "rbernoulli", p=0.3) +
+    node_td("C", "time_to_event", prob_fun=prob_fun)
+
+  out <- sim_n_datasets(dag=dag, n_sim=100, n_repeats=2, max_t=10,
+                        n_cores=2)
+})
+
 }